apps/market-service/pyproject.toml
# Project configuration
[project]
name = "market-service"
version = "0.1.0"
description = "AegisSDK service: market-service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "httpx>=0.24.0",
    "python-dateutil>=2.8.0",
    "clickhouse-driver>=0.2.6",
    "vnpy>=4.1.0",
    # Note: aegis-sdk will be installed from local copy in Docker
    "aegis-sdk",
    "testcontainers[nats]>=4.12.0",
    "meson-python>=0.18.0",
    "meson>=1.8.3",
    "pybind11>=3.0.0",
    "ninja>=1.13.0",
    "vnpy-sopt>=3.7.1.0",
    "importlib-metadata>=8.7.0",
    "vnpy-ctp>=6.7.7.2",
    "psutil>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["market-service"]

[tool.ruff]
line-length = 100
target-version = "py313"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --cov=market-service --cov-report=term-missing"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
markers = [
    "real_ctp: marks tests that require real CTP account credentials"
]

[tool.uv]
dev-dependencies = [
    "ipython>=8.0.0",
    "ipdb>=0.13.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-asyncio>=1.1.0",
    "black>=25.1.0",
    "ruff>=0.12.8",
    "mypy>=1.17.1",
    "packaging>=25.0",
]

[tool.uv.sources]
aegis-sdk = { path = "../../packages/aegis-sdk", editable = true }
